estimate time spent on each task
This tracks the time spent on each task with the assumption that the
last focused task before the application exits is the primary task which
the user was working on since the last time polar notes was run. The
unix time of the application exit is stored in the state file so that
each time polar tales exits, it can compute the time between the current
exit and previous exit, and add that duration to the relevant task.

This time is also copied to the clipboard summary of notes so that the
user doesn't have to dig into the state file to get that information.