implement select and exit command
The select and exit command allows the user to assign the time estimate
to a task and immediately exit, without modifying the notes for the
task. This reduces the number of keystrokes for this scenario from four
(eg `e 0 ctrl-s`) down to just 2 (eg `s 0`). It also removes the control
key which is often a little harder to reach.