add ctrl-c command in edit mode
In edit mode, one can now press ctrl-c in addition to ctrl-s to save and
exit, with ctrl-c copying the note contents to clipboard and ctrl-s not.
Additionally, the notes are now *only* copied to the clipboard by using
ctrl-c in edit mode in order to prevent polar tales accidentally writing
over existing clipboard contents and potentially resulting in data loss
for the user.