add a one-second delay before handling messages
Applications that steal focus are already annoying enough as is, so this
should help mitigate some of the annoyance by refusing to handle
keypresses until one second of time has passed. This way if a user is
in the middle of typing when focus is stolen, they don't start inputting
a bunch of commands they did not intend to.