implement time tracking for `x` command
This adds a field to the save state so that the most recently used log
entry can be tracked across program runs. This allows the `x` command to
infer that the user was working on the same task as last time the
program exited but had no further notes to add. In many cases this means
the user can reap the time tracking benefits while only having to press
a single key to exit the application and get back to real work.