include serde_derive in Cargo.toml
Windows does not seem to come with the derive feature enabled
automatically so including serde_derive is needed on windows. It is not
problematic to include it on linux so special cases for each OS did not
seem necessary. Current recommendations seem to be to use serde_derive
directly rather than enabling the feature in Cargo.toml.